Stop word lists should be sorted.
Stop word filters use a lunr.SortedSet which requires its elements to be
maintained in a sorted array. This change ensures that the stop words
used to load the sorted set are already sorted.

If the stop words are not presorted then the stop word filter will not
filter all the required stop words.