Upgrade to support lunr 2.x
Lunr 2.x has changed the interface for pipeline functions, instead of
just passing around strings, tokens are now represented as an instance
of lunr.Token.

Additionally, Lunr now supports per index tokenisers, this allows the JP
plugin to not have to clobber the global lunr.tokenizer function. This
is fine for when building the index, but queries are parsed by
lunr.QueryParser which may need some help when it comes to parsing
Japanese searches. Currently there is no extension point for the query
parser so we may need to do some work here.

This commit also includes updates for the demos.