Support beautifying and minifying hand crafted langs
lunr.ja.js is currently hand-crafted and as such, it is not picked
up during the build process. This causes two issues:

1. we don't get minified versions

2. code style is not consitent with generated lang files (i.e. indentation)

The updates in this patch allow us to perform beautification and
minification on language files that are hand crafted by simply adding
an entry to the file list with locale only.

Note: 'jp' was added to keep things consistent. It can be removed when
lunr.jp.js is removed.

Fixes #34